hypertext/requirements.txt
chevron
python-dateutil